test(server): separate mpe add tracks and delete tracks tests
