+++ packages/server/tests/MpeAddTracks.test.ts
@@ -2,7 +2,6 @@ import Database from '@ioc:Adonis/Lucid/Database';
 import {
     AllServerToClientEvents,
     MpeAcknowledgeAddingTracksRequestBody,
-    MpeAcknowledgeDeletingTracksRequestBody,
     MpeRejectAddingTracksRequestBody,
 } from '@musicroom/types';
 import MpeServerToTemporalController from 'App/Controllers/Http/Temporal/MpeServerToTemporalController';
@@ -16,14 +15,13 @@ import {
     BASE_URL,
     initTestUtils,
     generateMpeWorkflowState,
-    createSpyOnClientSocketEvent,
 } from './utils/TestUtils';
 
 function noop() {
     return undefined;
 }
 
-test.group('MPE Rooms Tracks List Editing', (group) => {
+test.group('MPE Add Tracks', (group) => {
     const {
         createUserAndGetSocket,
         createSocketConnection,
@@ -392,226 +390,5 @@ test.group('MPE Rooms Tracks List Editing', (group) => {
         });
 
         assert.isTrue(userBSocket1AddTracksSuccessCallbackSpy.notCalled);
-    });
-
-    test('Sends acknowledgement to the user if deleting tracks succeeded', async (assert) => {
-        const creatorUserID = datatype.uuid();
-        const roomID = datatype.uuid();
-        const userASocket1 = await createUserAndGetSocket({
-            userID: creatorUserID,
-            mpeRoomIDToAssociate: [
-                {
-                    roomID,
-                },
-            ],
-        });
-        const roomState = generateMpeWorkflowState({
-            roomID,
-            roomCreatorUserID: creatorUserID,
-        });
-
-        sinon
-            .stub(MpeServerToTemporalController, 'deleteTracks')
-            .callsFake(async ({ deviceID, userID }) => {
-                const body: MpeAcknowledgeDeletingTracksRequestBody = {
-                    deviceID,
-                    userID,
-                    state: {
-                        ...roomState,
-                        tracks: [],
-                    },
-                };
-
-                await supertest(BASE_URL)
-                    .post(
-                        urlcat(
-                            MPE_TEMPORAL_LISTENER,
-                            'acknowledge-deleting-tracks',
-                        ),
-                    )
-                    .send(body)
-                    .expect(200);
-
-                return Promise.resolve({
-                    ok: 1,
-                });
-            });
-
-        const userASocket1DeleteTracksSuccessCallbackSpy =
-            createSpyOnClientSocketEvent(
-                userASocket1,
-                'MPE_DELETE_TRACKS_SUCCESS_CALLBACK',
-            );
-        const userASocket1TracksListUpdateSpy = createSpyOnClientSocketEvent(
-            userASocket1,
-            'MPE_TRACKS_LIST_UPDATE',
-        );
-
-        const tracksToDelete = roomState.tracks.map((track) => track.id);
-        userASocket1.emit('MPE_DELETE_TRACKS', {
-            roomID,
-            tracksIDs: tracksToDelete,
-        });
-
-        await waitFor(() => {
-            assert.isTrue(
-                userASocket1DeleteTracksSuccessCallbackSpy.calledOnceWithExactly(
-                    {
-                        roomID,
-                        state: {
-                            ...roomState,
-                            tracks: [],
-                        },
-                    },
-                ),
-            );
-        });
-
-        assert.isTrue(userASocket1TracksListUpdateSpy.notCalled);
-    });
-
-    test("Sends updated tracks list to other user's devices if deleting tracks succeeded", async (assert) => {
-        const creatorUserID = datatype.uuid();
-        const roomID = datatype.uuid();
-        const userASocket1 = await createUserAndGetSocket({
-            userID: creatorUserID,
-            mpeRoomIDToAssociate: [
-                {
-                    roomID,
-                },
-            ],
-        });
-        const userASocket2 = await createSocketConnection({
-            userID: creatorUserID,
-        });
-        const roomState = generateMpeWorkflowState({
-            roomID,
-            roomCreatorUserID: creatorUserID,
-        });
-
-        sinon
-            .stub(MpeServerToTemporalController, 'deleteTracks')
-            .callsFake(async ({ deviceID, userID }) => {
-                const body: MpeAcknowledgeDeletingTracksRequestBody = {
-                    deviceID,
-                    userID,
-                    state: {
-                        ...roomState,
-                        tracks: [],
-                    },
-                };
-
-                await supertest(BASE_URL)
-                    .post(
-                        urlcat(
-                            MPE_TEMPORAL_LISTENER,
-                            'acknowledge-deleting-tracks',
-                        ),
-                    )
-                    .send(body)
-                    .expect(200);
-
-                return Promise.resolve({
-                    ok: 1,
-                });
-            });
-
-        const userASocket2TracksListUpdateSpy = createSpyOnClientSocketEvent(
-            userASocket2,
-            'MPE_TRACKS_LIST_UPDATE',
-        );
-
-        const tracksToDelete = roomState.tracks.map((track) => track.id);
-        userASocket1.emit('MPE_DELETE_TRACKS', {
-            roomID,
-            tracksIDs: tracksToDelete,
-        });
-
-        await waitFor(() => {
-            assert.isTrue(
-                userASocket2TracksListUpdateSpy.calledOnceWithExactly({
-                    roomID,
-                    state: {
-                        ...roomState,
-                        tracks: [],
-                    },
-                }),
-            );
-        });
-    });
-
-    test('Sends updated tracks list to all other users if deleting tracks succeeded', async (assert) => {
-        const creatorUserID = datatype.uuid();
-        const roomID = datatype.uuid();
-        const userASocket1 = await createUserAndGetSocket({
-            userID: creatorUserID,
-            mpeRoomIDToAssociate: [
-                {
-                    roomID,
-                },
-            ],
-        });
-        const userBSocket1 = await createUserAndGetSocket({
-            userID: datatype.uuid(),
-            mpeRoomIDToAssociate: [
-                {
-                    roomID,
-                },
-            ],
-        });
-        const roomState = generateMpeWorkflowState({
-            roomID,
-            roomCreatorUserID: creatorUserID,
-        });
-
-        sinon
-            .stub(MpeServerToTemporalController, 'deleteTracks')
-            .callsFake(async ({ deviceID, userID }) => {
-                const body: MpeAcknowledgeDeletingTracksRequestBody = {
-                    deviceID,
-                    userID,
-                    state: {
-                        ...roomState,
-                        tracks: [],
-                    },
-                };
-
-                await supertest(BASE_URL)
-                    .post(
-                        urlcat(
-                            MPE_TEMPORAL_LISTENER,
-                            'acknowledge-deleting-tracks',
-                        ),
-                    )
-                    .send(body)
-                    .expect(200);
-
-                return Promise.resolve({
-                    ok: 1,
-                });
-            });
-
-        const userBSocket1TracksListUpdateSpy = createSpyOnClientSocketEvent(
-            userBSocket1,
-            'MPE_TRACKS_LIST_UPDATE',
-        );
-
-        const tracksToDelete = roomState.tracks.map((track) => track.id);
-        userASocket1.emit('MPE_DELETE_TRACKS', {
-            roomID,
-            tracksIDs: tracksToDelete,
-        });
-
-        await waitFor(() => {
-            assert.isTrue(
-                userBSocket1TracksListUpdateSpy.calledOnceWithExactly({
-                    roomID,
-                    state: {
-                        ...roomState,
-                        tracks: [],
-                    },
-                }),
-            );
-        });
     });
 });
